requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
Flask-Migrate
Flask-WTF
flask-talisman
requests
scrypt
SQLAlchemy-Continuum
https://git.archlinux.org/pyalpm.git/snapshot/pyalpm-0.8.4.tar.gz
